HA suppport for eventing control plane
